refactor: 유저,인증,인가 DTO파일구조 개선 소셜로그인 에러코드 수정
