--- src/main/java/com/example/wherewego/domain/auth/Provider.java
@@ -1,5 +1,0 @@
-package com.example.wherewego.domain.auth;
-
-public enum Provider {
-	LOCAL, GOOGLE, KAKAO, NAVER
-}
