--- src/main/java/com/example/wherewego/domain/auth/UserRole.java
@@ -1,5 +1,0 @@
-package com.example.wherewego.domain.auth;
-
-public enum UserRole {
-	USER, ADMIN
-}
