+++ src/main/java/com/example/wherewego/domain/auth/dto/social/GoogleUserInfo.java
@@ -1,4 +1,4 @@
-package com.example.wherewego.domain.auth.dto.request;
+package com.example.wherewego.domain.auth.dto.social;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
