+++ src/main/java/com/example/wherewego/domain/auth/enums/Provider.java
@@ -1,0 +1,5 @@
+package com.example.wherewego.domain.auth.enums;
+
+public enum Provider {
+	LOCAL, GOOGLE, KAKAO
+}
