+++ src/main/java/com/example/wherewego/domain/auth/enums/UserRole.java
@@ -1,0 +1,5 @@
+package com.example.wherewego.domain.auth.enums;
+
+public enum UserRole {
+	USER, ADMIN
+}
