+++ src/main/java/com/example/wherewego/domain/user/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import com.example.wherewego.domain.auth.Provider;
+import com.example.wherewego.domain.auth.enums.Provider;
 import com.example.wherewego.domain.user.entity.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
